refactor fitworkspace classes
